refactored to display notification button on main menu component instead of incubator
